+++ CHANGELOG.md
@@ -7,3 +7,6 @@ All notable changes to this project will be documented in this file. This change
 Added
 * Updated dependencies
 * Added gh-pages and API docs
+* Re-implemented http-get using clj-http (as it is required by elastisch)
+* Fixed db/stop-db-exit behavior while in the REPL
+* Disabled web client tests for now